+++ src/index.ts
@@ -7,3 +7,5 @@ export type {
   InflectionChatModelId,
   InflectionChatSettings,
 } from "./inflection-chat-settings";
+export { generateObject } from "./inflection-generate-object";
+export type { GenerateObjectOptions } from "./inflection-generate-object";